refactor(fs.walk): get rid of __exportStar helper
