--- packages/fs/fs.walk/src/providers/index.ts
@@ -1,3 +1,0 @@
-export * from './async';
-export * from './stream';
-export * from './sync';
