--- packages/fs/fs.walk/src/readers/index.ts
@@ -1,2 +1,0 @@
-export * from './async';
-export * from './sync';
